C2-NAV.33 pre-registration: the gate, the one leaf, and the falsifier
Committed BEFORE the simulator. No live run in this commit; no
behavioural parameter is loaded by anything here.

THE GATE IS CLOSED FIRST, because the brief forbids assuming that
resample_interval: 2 exposes anything. Upstream amcl_node.cpp is NOT
installed -- only the headers and the .so -- so the proof is made from
what is actually on the machine, and by the module at run time rather
than pasted in.

libamcl_core.so has EXACTLY ONE pf_update_resample call site and EXACTLY
ONE publishParticleCloud call site in the whole library, both inside
laserReceived, and their guards differ. The resample is behind
`idivl 0xac8(%rbx)` on ++resample_count_ at +0x8e0; the publish is behind
`movzbl 0x5c8(%rbx)` -- force_update_ -- and NOT behind whether the
resample ran. The set handed to the publisher is
&pf_->sets[pf_->current_set], the address arithmetic checked against the
installed pf.hpp (current_set +0x18, sets +0x20, and the disassembly's
rax*9<<4 = 144 = sizeof(pf_sample_set_t) laid out from that header). The
weight is copied verbatim: movsd -0x8(%r15),%xmm0 ; movsd %xmm0,-0x8(%r14).
So on an update where the resample is skipped, current_set still indexes
the set pf_update_sensor just weighted.

c2nav33_oracle.cpp EXECUTES the deployed filter rather than describing
it: it links libpf_lib.so and calls pf_update_sensor and
pf_update_resample itself, and contains no particle-filter equations. At
n=500, after pf_update_sensor there are 500 distinct weights summing to 1
with ESS/n 0.7777; after pf_update_resample there is ONE bit-identical
weight at 1/500 and current_set has flipped 0 -> 1. That second line IS
C2-NAV.30's own measurement -- ESS/n = 1.0000 on 40 of 40 -- which makes
this a prediction with a matching record rather than a story.

ONE LEAF, over 323. paramdiff: added 0, removed 0, changed 1 --
amcl.ros__parameters.resample_interval 1 -> 2. All 38 guarded AMCL leaves
identical and present, 0 non-AMCL leaves changed. Baseline
c2nav25_slow_params.yaml sha256 4c15893e8393, candidate
c2nav33_ri2_params.yaml sha256 1f4e29a12619.

THE DISCRIMINATOR IS FIXED IN ADVANCE, and it is NOT south_w > north_w --
that would be moving the goalposts, because C2-NAV.30 measured only 36.4%
of particles north of GT, so a NEUTRAL weighting already gives
south_w ~ 0.64. The cloud's existing displacement is divided out:
shift = weighted dy - unweighted dy, and mass_shift = south_w - south_u,
both on the SAME particles. (A) requires the 95% CI of the mean shift to
lie wholly below zero; (B) is anything else given observed pre-resample
clouds; (C) is no pre-resample cloud at all.

AND THE FAILED-EXPERIMENT CASE IS DETECTED, NOT ASSUMED AWAY. A run whose
clouds are all bit-flat means the parameter did not take effect, which is
(C) and not "the weights are unbiased". Two independent detectors, per
the CLAUDE.md rule that a "we saw nothing" finding must come from an
instrument proven able to see something. c2nav33_liveparam.sh reads the
parameter off the RUNNING /amcl -- nothing in the existing runner reads
any amcl parameter, and initParameters carries a `movl $0x1,0xac8(%r12)`
reached through a WARN-severity logging block, i.e. a silent fallback to
1 that would look exactly like an unbiased result. And `phase` classifies
every cloud from the weights themselves by a bit-level all-equal test and
requires the phases to alternate.

FOUND ON THE WAY, NOT ASKED FOR. objdump emits a symbol's @plt trampoline
at a LOWER address than its definition, so "first label that matches"
returns a two-line stub -- which failed every content check and reported
the binary as not doing what it demonstrably does. A false NEGATIVE, in
the safe direction, but it is now skipped explicitly and the selftest
exercises it.

ONE INSTRUMENTATION CHANGE: nav_bench.py's cloud snapshots go float32 ->
float64. C2-NAV.30 asked cloud-GEOMETRY questions, where float32's
~0.24 um on a 2 m coordinate is far below anything it measured. This
session's measurement IS the weight field and the discriminator is
whether weights are EXACTLY flat, so storing the quantity under test at
reduced precision is the hidden filtering that would make a negative
result unreadable. No behavioural parameter and no other line of
nav_bench.py is touched.

selftest 59 checks, 0 failed, offline, before the simulator: every
arithmetic claim against a hand-computed answer, the null control (flat
weights must give exactly zero) on a hand-built flat cloud, the sign
checked on a MIRRORED fixture so a shift that came out negative
regardless would fail, missing data asserted None rather than filled, and
C2-NAV.28/.30/.31/.32 all shown to still import and reproduce.

Nothing here proposes resample_interval: 2 as a configuration. It is a
diagnostic window and it must be reverted.

+++ gazebo_models/scripts/nav_bench.py
@@ -558,7 +558,7 @@ class NavBench(Node):
         # above: resident only, and only the per-leg window is written
         # out. 2000 particles x 4 float32 = 32 KB, so a 1200-entry ring
         # is ~38 MB, of the same order as the costmap ring already here.
-        self.cloudsnaps = []        # (tsim, (n,4) float32), ring, last 1200
+        self.cloudsnaps = []        # (tsim, (n,4) float64), ring, last 1200
         self.cloud_frame = None     # header.frame_id, RECORDED not assumed
         self.n_cloud_msgs = 0
 
@@ -664,7 +664,15 @@ class NavBench(Node):
         same eigen-decomposition ten times for one message."""
         ts, tw = self.now()
         n = len(m.particles)
-        arr = np.empty((n, 4), dtype=np.float32)
+        # C2-NAV.33. float64, not float32. C2-NAV.30 asked cloud-GEOMETRY
+        # questions, where float32's ~0.24 um on a 2 m coordinate is far
+        # below anything it measured. This session's measurement IS the
+        # weight field, and the discriminator it turns on is whether the
+        # weights are EXACTLY flat (ESS/n == 1 to the last bit, which is
+        # what resampling produces) or merely nearly so. Storing the one
+        # quantity under test at reduced precision is the kind of hidden
+        # filtering that would make a negative result unreadable.
+        arr = np.empty((n, 4), dtype=np.float64)
         for i, part in enumerate(m.particles):
             pos, o = part.pose.position, part.pose.orientation
             arr[i] = (pos.x, pos.y, yaw_of(o), part.weight)
@@ -1908,7 +1916,7 @@ def main(argv=None):
             pc_base = os.path.join(tracedir, f'{name}_rep{rep}_cloud')
             counts = np.array([len(a) for (_, a) in snaps], dtype=np.int32)
             parts = (np.concatenate([a for (_, a) in snaps], axis=0)
-                     if snaps else np.zeros((0, 4), dtype=np.float32))
+                     if snaps else np.zeros((0, 4), dtype=np.float64))
             np.savez_compressed(
                 pc_base + '.npz', particles=parts, counts=counts,
                 ts_sim_s=np.array([ts for (ts, _) in snaps],
